move strtobool to `paddle.utils`
