+++ python/paddle/distributed/launch/context/args_envs.py
@@ -13,12 +13,9 @@
 # limitations under the License.
 
 import os
-import pathlib
-import sys
 from argparse import REMAINDER, ArgumentParser
 
-sys.path.append(str(pathlib.Path(__file__).resolve().parents[2] / 'utils'))
-from launch_utils import strtobool
+from paddle.utils import strtobool
 
 env_args_mapping = {
     'POD_IP': ('host', str),
