+++ python/paddle/distributed/utils/launch_utils.py
@@ -22,6 +22,7 @@ import time
 from contextlib import closing
 
 from paddle.distributed.fleet.launch_utils import get_backend_by_compile_flag
+from paddle.utils import strtobool
 
 from ..utils.log_utils import get_logger
 
@@ -551,13 +552,3 @@ def _print_arguments(args):
     for arg, value in sorted(vars(args).items()):
         print(f"{arg}: {value}")
     print("------------------------------------------------")
-
-
-def strtobool(val):
-    val = val.lower()
-    if val in ['y', 'yes', 't', 'true', 'on', '1']:
-        return True
-    elif val in ['n', 'no', 'f', 'false', 'off', '0']:
-        return False
-    else:
-        raise ValueError(f"Invalid truth value {val!r}")
